Doing more work on the battle infrastructure:
So far, the match simply consists of two players dealing damage to each other. I plan on adding a chat log for each player, which would be formatted
the same way as the data scraped from showdown's chat would. This way, we can simulate the same degree of uncertainty our agent will have in the first few rounds of the game against a real player by only displaying enemy pokemon as they're played and attack.